Import files from the Webmachine project
Webmachine is an application layer that adds HTTP semantic awareness
on top of the excellent bit-pushing and HTTP syntax-management
provided by Hunchentoot, and provides a simple and clean way to
connect that to your application's behavior.

The design is inspired by the mythic Webmachine of Erlang and honours
it with that name.